+++ pkg/service/wallet.go
@@ -6,14 +6,14 @@ import (
 	"github.com/google/uuid"
 )
 
-func (s *Service) Withdraw(ctx context.Context, walletID uuid.UUID, amount float64) error {
+func (s *Service) Withdraw(ctx context.Context, walletID uuid.UUID, amount int) error {
 	return s.Database.Withdraw(ctx, walletID, amount)
 }
 
-func (s *Service) Deposit(ctx context.Context, walletID uuid.UUID, amount float64) error {
+func (s *Service) Deposit(ctx context.Context, walletID uuid.UUID, amount int) error {
 	return s.Database.Deposit(ctx, walletID, amount)
 }
 
-func (s *Service) GetBalance(ctx context.Context, walletID uuid.UUID) (float64, error) {
+func (s *Service) GetBalance(ctx context.Context, walletID uuid.UUID) (int, error) {
 	return s.Database.GetBalance(ctx, walletID)
 }
